Completed the Exercise 10 - Return a specific Fibonacci number
I solved the exercise using both - iterative and recursive way. What
was interesting for me to find out is that the recursive way of doing it
takes considerably more time to complete as the number gets bigger. I
read that this is due to the time complexity of recursion which is
grows exponentially while the iterative way has a linear time complexity.